+++ R/backend-xla.R
@@ -367,8 +367,12 @@ xla <- function(f, args, donate = character(), device = NULL) {
 #' @export
 AnvlBackendXla <- function() {
   backend <- AnvlBackend(
-    new_data = function(data, dtype, shape, device, ambiguous) {
-      buf <- pjrt_buffer(data, dtype = dtype, device = device, shape = shape)
+    new_data = function(data, dtype, shape, device, ambiguous, row_major = FALSE) {
+      buf <- if (is.raw(data)) {
+        pjrt_buffer(data, dtype = dtype, device = device, shape = shape, row_major = row_major)
+      } else {
+        pjrt_buffer(data, dtype = dtype, device = device, shape = shape)
+      }
       structure(
         list(data = buf, ambiguous = ambiguous, backend = "xla"),
         class = "AnvlArray"